refactored transfer command
